Medical/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
langchain==0.1.12
langchain-openai==0.0.8
faiss-cpu==1.7.4; platform_system != "Darwin"
faiss-cpu==1.7.4; platform_system == "Darwin" and platform_machine == "x86_64"
scikit-learn==1.4.1.post1; platform_system == "Darwin" and platform_machine == "arm64"
ragas==0.1.3
pymupdf==1.23.26
pydantic==2.6.4
prometheus-client==0.20.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
pytest==8.0.2
httpx==0.27.0 